新增：模板文件上传接口

+++ sz-service/sz-service-admin/src/main/java/com/sz/admin/system/pojo/po/SysTempFile.java
@@ -9,13 +9,13 @@ import com.sz.mysql.EntityChangeListener;
 import java.time.LocalDateTime;
 
 /**
-* <p>
-* 模版文件表
-* </p>
-*
-* @author sz-admin
-* @since 2024-12-05
-*/
+ * <p>
+ * 模版文件表
+ * </p>
+ *
+ * @author sz-admin
+ * @since 2024-12-05
+ */
 @Data
 @Table(value = "sys_temp_file", onInsert = EntityChangeListener.class, onUpdate = EntityChangeListener.class)
 @Schema(description = "模版文件表")
@@ -25,35 +25,35 @@ public class SysTempFile implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id(keyType = KeyType.Auto)
-    @Schema(description ="")
+    @Schema(description = "")
     private Long id;
 
-    @Schema(description ="文件ID")
-    private Integer sysFileId;
+    @Schema(description = "文件ID")
+    private Long sysFileId;
 
-    @Schema(description ="模版名")
+    @Schema(description = "模版名")
     private String tempName;
 
-    @Schema(description ="地址")
+    @Schema(description = "地址")
     private String url;
 
-    @Schema(description ="备注")
+    @Schema(description = "备注")
     private String remark;
 
     @Column(isLogicDelete = true)
-    @Schema(description ="逻辑删除")
+    @Schema(description = "逻辑删除")
     private String delFlag;
 
-    @Schema(description ="创建人")
+    @Schema(description = "创建人")
     private Long createId;
 
-    @Schema(description ="创建时间")
+    @Schema(description = "创建时间")
     private LocalDateTime createTime;
 
-    @Schema(description ="更新人")
+    @Schema(description = "更新人")
     private Long updateId;
 
-    @Schema(description ="更新时间")
+    @Schema(description = "更新时间")
     private LocalDateTime updateTime;
 
 }
